daemon: extract dataplane+HA core apply from daemon_apply.go
Pure code-motion split of pkg/daemon/daemon_apply.go (#5661): the
dataplane/HA core apply phase (networkd apply, userspace dataplane
program, cluster wiring) and the deferred-worker-arm bookkeeping move
verbatim into daemon_apply_dataplane.go. Apply ordering and side-effect
sequence inside applyDataplaneAndHACore are unchanged.

Moved declarations (byte-identical bodies, no rename, no logic change):
applyDataplaneAndHACore, setDataplaneDeferWorkers,
reapplyAfterDeferredMAC, recordDataplaneWorkerArmDebt. Same package, no
caller changes; per-file imports pruned by goimports. Build: go build
./pkg/daemon/... green.
